Initial draft for class Operators and Recording on TriMesh
Notable changes:
* Each Operation (and its before/after checks) are derived from TriMeshOperator and their full execution is handled by the class
* TriMesh topology (vertices and triangles) are now Attributes, which moves the burden of undoing operations from the operator to the Attribute class

Initial implementation was thanks to zlyfunction and adapted by mtao.
To pursue the history in more detail, changes come from mtao/wildmeshing-toolkit:mtao/TriMesh_event_logging (around 3e3d7c0), which comes from zlyfunction/wildmeshing-toolkit:main

+++ tests/2d_tests.cpp
@@ -1,10 +1,27 @@
+
 #include <wmtk/TriMesh.h>
+
 
 #include <igl/read_triangle_mesh.h>
 #include <stdlib.h>
 #include <catch2/catch.hpp>
+#include <highfive/H5File.hpp>
 #include <iostream>
 
+
+// template <>
+// HighFive::DataType HighFive::create_datatype<TriMesh::VertexConnectivity>() {
+// }
+// template <>
+// HighFive::DataType HighFive::create_datatype<TriMesh::TriangleConnectivity>(){
+//    {"indices", HighFive::create_datatype<size_t[3]>()},
+//    {"is_removed", HighFive::create_datatype<bool>()},
+//    {"hash", HighFive::create_datatype<size_t>()},
+
+//}
+// WMTK_HDF5_REGISTER_ATTRIBUTE_TYPE(wmtk::TriMesh::VertexConnectivity)
+// WMTK_HDF5_REGISTER_ATTRIBUTE_TYPE(wmtk::TriMesh::TriangleConnectivity)
+//
 using namespace wmtk;
 
 TEST_CASE("load mesh and create TriMesh", "[test_mesh_creation]")
@@ -111,9 +128,9 @@ TEST_CASE("random 10 switches on 2 traingles", "[tuple_operation]")
     SECTION("test all tuples generated using vertices")
     {
         auto vertices_tuples = m.get_vertices();
-        for (int i = 0; i < vertices_tuples.size(); i++) {
+        for (size_t i = 0; i < vertices_tuples.size(); i++) {
             TriMesh::Tuple v_tuple = vertices_tuples[i];
-            for (int j = 0; j < 10; j++) {
+            for (size_t j = 0; j < 10; j++) {
                 size_t test = rand() % 3;
                 switch (test) {
                 case 0: v_tuple = v_tuple.switch_vertex(m); break;
@@ -129,9 +146,9 @@ TEST_CASE("random 10 switches on 2 traingles", "[tuple_operation]")
     SECTION("test all tuples generated using edges")
     {
         auto edges_tuples = m.get_edges();
-        for (int i = 0; i < edges_tuples.size(); i++) {
+        for (size_t i = 0; i < edges_tuples.size(); i++) {
             TriMesh::Tuple e_tuple = edges_tuples[i];
-            for (int j = 0; j < 10; j++) {
+            for (size_t j = 0; j < 10; j++) {
                 size_t test = rand() % 3;
                 switch (test) {
                 case 0: e_tuple = e_tuple.switch_vertex(m); break;
@@ -147,9 +164,9 @@ TEST_CASE("random 10 switches on 2 traingles", "[tuple_operation]")
     SECTION("test all tuples generated using faces")
     {
         auto faces_tuples = m.get_faces();
-        for (int i = 0; i < faces_tuples.size(); i++) {
+        for (size_t i = 0; i < faces_tuples.size(); i++) {
             TriMesh::Tuple f_tuple = faces_tuples[i];
-            for (int j = 0; j < 10; j++) {
+            for (size_t j = 0; j < 10; j++) {
                 size_t test = rand() % 3;
                 switch (test) {
                 case 0: f_tuple = f_tuple.switch_vertex(m); break;
@@ -204,7 +221,7 @@ TEST_CASE("double switches is identity", "[tuple_operation]")
     {
         TriMesh::Tuple v_tuple_after;
         auto vertices_tuples = m.get_vertices();
-        for (int i = 0; i < vertices_tuples.size(); i++) {
+        for (size_t i = 0; i < vertices_tuples.size(); i++) {
             TriMesh::Tuple v_tuple = vertices_tuples[i];
             v_tuple_after = double_switch_vertex(v_tuple, m);
             REQUIRE(tuple_equal(m, v_tuple_after, v_tuple));
@@ -219,7 +236,7 @@ TEST_CASE("double switches is identity", "[tuple_operation]")
     {
         TriMesh::Tuple e_tuple_after;
         auto edges_tuples = m.get_edges();
-        for (int i = 0; i < edges_tuples.size(); i++) {
+        for (size_t i = 0; i < edges_tuples.size(); i++) {
             TriMesh::Tuple e_tuple = edges_tuples[i];
             e_tuple_after = double_switch_vertex(e_tuple, m);
             REQUIRE(tuple_equal(m, e_tuple_after, e_tuple));
@@ -234,7 +251,7 @@ TEST_CASE("double switches is identity", "[tuple_operation]")
     {
         TriMesh::Tuple f_tuple_after;
         auto faces_tuples = m.get_faces();
-        for (int i = 0; i < faces_tuples.size(); i++) {
+        for (size_t i = 0; i < faces_tuples.size(); i++) {
             TriMesh::Tuple f_tuple = faces_tuples[i];
             f_tuple_after = double_switch_vertex(f_tuple, m);
             REQUIRE(tuple_equal(m, f_tuple_after, f_tuple));
@@ -257,7 +274,7 @@ TEST_CASE("vertex_edge switches equals indentity", "[tuple_operation]")
     {
         TriMesh::Tuple v_tuple_after;
         auto vertices_tuples = m.get_vertices();
-        for (int i = 0; i < vertices_tuples.size(); i++) {
+        for (size_t i = 0; i < vertices_tuples.size(); i++) {
             TriMesh::Tuple v_tuple = vertices_tuples[i];
             v_tuple_after = v_tuple.switch_vertex(m);
             v_tuple_after = v_tuple_after.switch_edge(m);
@@ -273,7 +290,7 @@ TEST_CASE("vertex_edge switches equals indentity", "[tuple_operation]")
     {
         TriMesh::Tuple e_tuple_after;
         auto edges_tuples = m.get_edges();
-        for (int i = 0; i < edges_tuples.size(); i++) {
+        for (size_t i = 0; i < edges_tuples.size(); i++) {
             TriMesh::Tuple e_tuple = edges_tuples[i];
             e_tuple_after = e_tuple.switch_vertex(m);
             e_tuple_after = e_tuple_after.switch_edge(m);
@@ -289,7 +306,7 @@ TEST_CASE("vertex_edge switches equals indentity", "[tuple_operation]")
     {
         TriMesh::Tuple f_tuple_after;
         auto faces_tuples = m.get_faces();
-        for (int i = 0; i < faces_tuples.size(); i++) {
+        for (size_t i = 0; i < faces_tuples.size(); i++) {
             TriMesh::Tuple f_tuple = faces_tuples[i];
             f_tuple_after = f_tuple.switch_vertex(m);
             f_tuple_after = f_tuple_after.switch_edge(m);
@@ -485,3 +502,5 @@ TEST_CASE("split_operation", "[test_2d_operation]")
         for (auto e : edges) REQUIRE_FALSE(e.is_valid(m));
     }
 }
+
+
